basic login authentication was added, looking to enhance by applying JWT token-based access
